# docs/swagger.yaml
definitions:
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      errors: {}
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is a docs for e-commerce API.
  title: E-Commerce API
paths:
...
  /public/users/login:
    post:
      operationId: Login
      parameters:
      - description: email
        in: formData
        name: email
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.ErrorResponse'
            - properties:
                errors:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.ErrorResponse'
            - properties:
                errors:
                  items:
                    type: string
                  type: array
              type: object
      summary: Login User
      tags:
      - User
...
securityDefinitions:
  JWTAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"